requirements.txt
pdfplumber
python-docx
customtkinter
